perf!: Use Maps to store all players instead of an array
Quite frequently in the Skribbler Client, we have to loop over the entire players array every time an action related to a player occurs, such as when a player joins the lobby, sends a chat message, guesses a word, etc. Although vanilla Skribbl game servers only allow a maximum of 20 players which is not much to loop over, custom or modded servers can have any arbitrary number of players online in a single lobby. To ensure that the Skribbler Client can effectively handle a large player count in the thousands, we now use Maps to store the player list which allows for much faster item lookup times compared to looping over an array. This change does have a drawback though, and that is players with duplicated IDs will override one another. In this case, a custom server owner should update their server to avoid player-ID collisions which would have many bugs regardless of this.

A benchmark of array lookups with .find\(\) and map lookups with .get\(\) with 999999 items can be found at https://jsbench.me/qwmj3p1ou3/1